Feat: [#12] 팔로우홈피드셋팅

+++ back/dipping/src/main/java/com/common/dipping/repository/BoardSongRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.common.dipping.domain.entity.BoardSong;
 
-public interface BoardSongRepository extends JpaRepository<BoardSong, Integer>{
+public interface BoardSongRepository extends JpaRepository<BoardSong, Long>{
 
 }
