+++ back/dipping/src/main/java/com/common/dipping/repository/LikeRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.common.dipping.domain.entity.Like;
 
-public interface LikeRepository extends JpaRepository<Like, Integer>{
+public interface LikeRepository extends JpaRepository<Like, Long>{
 
 }
