+++ back/dipping/src/main/java/com/common/dipping/repository/PostTagRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.common.dipping.domain.entity.PostTag;
 
-public interface PostTagRepository extends JpaRepository<PostTag, Integer>{
+public interface PostTagRepository extends JpaRepository<PostTag, Long>{
 
 }
